backend/requirements.txt
google-genai>=1.14.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
Pillow>=10.4.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
